refactor: clean code and improve structure
